feat : BookingService 이미 예약된 좌석 요청 실패/ 좌석 예매 요청 성공 테스트 코드 작성

+++ src/main/java/com/example/demo/controller/BookingController.java
@@ -38,7 +38,7 @@ public class BookingController {
     }
 
     @GetMapping("/{movieId}/screenings")
-    public ResponseEntity<?> getScreeningsByMovie(@PathVariable("movieId") Long movieId){
+    public ResponseEntity<?> getScreeningsByMovie(@PathVariable("movieId") Integer movieId){
         try{
             return ResponseEntity.ok(movieScheduleService.getMovieWithSchedules(movieId));
         } catch(Exception e){
