+++ src/main/java/com/example/demo/service/MovieScheduleService.java
@@ -15,11 +15,11 @@ public class MovieScheduleService {
     private final ScreeningService screeningService;
     private final MovieScheduleRepository movieScheduleRepository;
 
-    public List<Screening> getScheduleByMovieId(Long movieId){
+    public List<Screening> getScheduleByMovieId(Integer movieId){
         return screeningService.getScreeningListByMovieId(movieId);
     }
 
-    public List<MovieScheduleDto> getMovieWithSchedules(Long movieId){
+    public List<MovieScheduleDto> getMovieWithSchedules(Integer movieId){
         return movieScheduleRepository.findByMovieId(movieId)
                 .orElseThrow(() -> new EntityNotFoundException("상영정보 없음"));
     }
