+++ src/main/java/com/example/demo/service/ScreeningService.java
@@ -30,7 +30,7 @@ public class ScreeningService {
             throw new EntityNotFoundException(String.format("해당 시간대에 영화 없음"));
     }
 
-    public List<Screening> getScreeningListByMovieId(Long movieId){
+    public List<Screening> getScreeningListByMovieId(Integer movieId){
         return screeningJpaRepository.findAllByMovieId(movieId)
                 .orElseThrow(() -> new EntityNotFoundException(String.format("상영 정보 없음")));
     }
